input descriptor is correctly created as long as the object is no deeper than 2 levels
